Fix two incompatible-pointer-types warnings

+++ bindings/elixir/native/src/ckzg_wrap.c
@@ -561,14 +561,14 @@ static ERL_NIF_TERM recover_cells_and_kzg_proofs_nif(
 
     ERL_NIF_TERM head;
     ERL_NIF_TERM tail = argv[0];
-    uint64_t current_u;
+    unsigned long current_u;
 
     // Check every cell index is an integer then store as native C type.
     for (int i = 0; enif_get_list_cell(env, tail, &head, &tail); i++) {
         if (!enif_get_uint64(env, head, &current_u))
             return make_error(env, ckzg_atoms.cell_indices_value_not_uint64);
 
-        cell_indices[i] = current_u;
+        cell_indices[i] = (uint64_t)current_u;
     }
 
     Cell *cells = enif_alloc(cells_len * BYTES_PER_CELL);
@@ -732,14 +732,14 @@ static ERL_NIF_TERM verify_cell_kzg_proof_batch_nif(
     tail = argv[1];
     // Check every cell index is an integer then store as native C type.
     for (int i = 0; enif_get_list_cell(env, tail, &head, &tail); i++) {
-        uint64_t current_u;
+        unsigned long current_u;
         if (!enif_get_uint64(env, head, &current_u)) {
             enif_free(commitments);
             enif_free(cell_indices);
             return make_error(env, ckzg_atoms.cell_indices_value_not_uint64);
         }
 
-        cell_indices[i] = current_u;
+        cell_indices[i] = (uint64_t)current_u;
     }
 
     Cell *cells = enif_alloc(cells_len * BYTES_PER_CELL);
